[Fix] buildingData undefined in Browser and Simulator
It turns out that when playing in browser and simulator, buildingData was always undefined when ToweBuilding tried to get its data from getBuildingDataById().

With this fix, a promise was implemented so that the code execution waits until the game settings data is successfully loaded from the Jsons files, allowing the project to run correctly in browser and simulator modes.