[Add] Multiple hire towers support
The building panel now works based on the selected tower building regardless of how many towers there are.

For now, in order to get the building panel to work properly, it is required that every tower has an unique (and non repeatable) tower id from the buildings.json file.

For this project, I added 4 additional towers that can hire different heroes based on their element.